Move publishing settings to a separate object
